fixed printing long address

+++ include/arm-gen.h
@@ -26,6 +26,6 @@
 
 // unsigned int, char *, size_t, int, int, size_t
 // Generate all the ARM gadgets
-gadget_list * generate_arm(unsigned int vma, char * rawbuf, size_t size, int bits, int endian, size_t depth);
+gadget_list * generate_arm(unsigned long long vma, char * rawbuf, size_t size, int bits, int endian, size_t depth);
 
 #endif
